+++ Sources/HIGThemesContract/HIGTheme+Defaults.swift
@@ -42,6 +42,7 @@ extension HIGTheme {
     public var chatBubble: any HIGChatBubbleTokens { HIGSystemChatBubbleTokens() }
     public var comment: any HIGCommentTokens { HIGSystemCommentTokens() }
     public var cover: any HIGCoverTokens { HIGSystemCoverTokens() }
+    public var imageOverlay: any HIGImageOverlayTokens { HIGSystemImageOverlayTokens() }
     public var dataTable: any HIGDataTableTokens { HIGSystemDataTableTokens() }
     public var breadcrumb: any HIGBreadcrumbTokens { HIGSystemBreadcrumbTokens() }
     public var pageHeader: any HIGPageHeaderTokens { HIGSystemPageHeaderTokens() }