+++ Sources/HIGThemesSystem/HIGSystemTheme.swift
@@ -41,6 +41,7 @@ public struct HIGSystemTheme: HIGTheme, Sendable {
     public let chatBubble: any HIGChatBubbleTokens
     public let comment: any HIGCommentTokens
     public let cover: any HIGCoverTokens
+    public let imageOverlay: any HIGImageOverlayTokens
     public let dataTable: any HIGDataTableTokens
     public let breadcrumb: any HIGBreadcrumbTokens
     public let pageHeader: any HIGPageHeaderTokens
@@ -130,6 +131,7 @@ public struct HIGSystemTheme: HIGTheme, Sendable {
         chatBubble: any HIGChatBubbleTokens = HIGSystemChatBubbleTokens(),
         comment: any HIGCommentTokens = HIGSystemCommentTokens(),
         cover: any HIGCoverTokens = HIGSystemCoverTokens(),
+        imageOverlay: any HIGImageOverlayTokens = HIGSystemImageOverlayTokens(),
         dataTable: any HIGDataTableTokens = HIGSystemDataTableTokens(),
         breadcrumb: any HIGBreadcrumbTokens = HIGSystemBreadcrumbTokens(),
         pageHeader: any HIGPageHeaderTokens = HIGSystemPageHeaderTokens(),
@@ -218,6 +220,7 @@ public struct HIGSystemTheme: HIGTheme, Sendable {
         self.chatBubble = chatBubble
         self.comment = comment
         self.cover = cover
+        self.imageOverlay = imageOverlay
         self.dataTable = dataTable
         self.breadcrumb = breadcrumb
         self.pageHeader = pageHeader